Bug#37190332 openid-connect fails if connection comes from pool
When a

1. the default-authentication-plugin set to openid-connect and
2. the connection-pool of the router is enabled
3. and connection is in the pool that can be reused

authentication currently fails with:

- Error 1, could not open id-token-file

Background
==========

The client plugin gets confused as it sent the token already in the
initial handshake (default-authentication-plugin=openid-connect) and
gets asked to switch to the same plugin. The id-token-file was already
used for the first round, and the 2nd round would need a new
id-token-file set.

Change
======

- remember the openid-connect token in the client-greeting to use it in
  the change-user that's sent to the server after taking the connection
  from the pool.

Change-Id: I925de99c790e7acffb5c84623890a5a48dec0125