Bug#37184880 Test main.opt_hint_resource_group_hypergraph failing testcase check Bug#37184884 Test main.opt_hints_join_order_hypergraph failing with a result diff
1) opt_hint_resource_group_hypergraph failed because the optimizer
switch was set to enable hypergraph, but never set to disable it.
This is because the test exited after enabling hypergraph and then
detecting that the test is disabled on Mac. This is fixed by
shifting the check for Mac platforms before the directive to
enable hypergraph.

2) opt_hints_join_order_hypergraph failed because one query had a
different EXPLAIN plan on Mac. This is fixed by modifying the
failing query to make the join order hint more restrictive. This
eliminates EXPLAIN plan variations across platforms by reducing
the possible join orders.

Change-Id: Iaa3188097ecc6fdab910f65bf687cac6e6c52398