Bug#37161583: Incorrect result with LEFT JOIN and small join_buffer_size [1/2, refactor, noclose]
Initial refactoring of the code.

The function WriteRowsToChunks() in hash_join_iterator.cc is written
as if it could be called either on the build input or on the probe
input. But it is only ever called on the build input.

Rename it to WriteBuildTableToChunkFiles(), make it a member function
of HashJoinIterator, and make it access the build input data
structures directly instead of taking the input as function arguments.

This makes the purpose of the function clearer, and it makes it easier
to see where it reads data from.

Change-Id: I09d7f23a6051597ac94f01d447e17e64761397be