Bug#37278449: Backport fix for bug #37161583 to MySQL Server 8.0/8.4 Bug#37161583: Incorrect result with LEFT JOIN and small join_buffer_size
Wrong results were returned for some queries that used hash antijoin
when the hash table didn't fit in the join buffer and spilled to
disk. (The bug title says LEFT JOIN, but the bug is actually in the
antijoin handling. The query in the bug report used LEFT JOIN, but it
was transformed internally from a left outer join to an antijoin.)

The problem was that some rows in the probe table were skipped when
writing the probe rows to chunk files. The skipped rows were those
that had a NULL value in a part of the join key. Such rows are fine to
skip for inner joins and semijoins, as they are known to have no match
in the build table. For outer joins and antijoins, however, the rows
in the probe table that have no matching row in the build table,
should be part of the join result, so they must be included in the
chunk files.

There was already logic for preserving these rows in the chunk files
for outer joins. This patch extends the logic so that it also covers
antijoins.

Change-Id: Iee4ab58d549b0c653a893ec18542adbf6130c4d6
(cherry picked from commit 01c8e25ddc441f3df4169352b5c4b76484043690)