Bug#37125935 COPY_FRAGREQ not taking fragment scan lock
COPY_FRAGREQ is used to synchronise starting data nodes
with a live node's table state.

Each fragment is scanned on a live node, generating
LQHKEYREQ signals towards the starting node.

This scan is performed by the fragment's owning LDM,
but can be concurrent with scans in Query threads on
the live node and so should be protected by fragment
locks.

Change-Id: Ibf9f20624cd50bfb40553a4d9a4324dad7f6ef8e