Bug#37143289: Missing error handling in Item_bool_func2::resolve_type()
We discovered that Item_bool_func2::resolve_type() makes an unchecked
call to Item_bool_func::resolve_type(), i.e. the call ignores the return
value and continues even in the case of an error. By convention, if we
get an error we want to stop whatever we are doing and allow it to
bubble up through the call stack. This patch wraps the inner call to
resolve_type() so that it conforms to this convention and returns true
early in the case of an error.

Change-Id: Iadaa45f2dc81b2dffb22475d47c8a24144fe5d1d