Bug#35660890 Scans of multi ranges can lose fragment locks for second and onwards ranges
Part 1/3
Add mechanism to check that fragment locks are released by
LQH/QLQH block executing threads after each signal execution
from the job buffer.

This gives a safety net around making changes to fragment
locking, checking that fragment lock tenure is always contained
within a single signal execution's scope.

The checks are added to the existing 'debug' global variable
initialisation, so not active in production.

Change-Id: If50e89ad35fb8a57f1500c085054b2199bcb1eb6

Bug#35660890 Scans of multi ranges can lose fragment locks for second and onwards ranges

Part 2/3

Add testcase showing the reported problem.

Reproducing the problem as-is requires concurrency between MRR
scans using query threads and update operations committing on
the LDM thread.

However the root cause of the failure is a lack of lock coverage
during the MRR scan.

Adding a self-check to the scan code that the required lock is
held causes the problem to surface simply by executing an
MRR scan on a query thread.

This patch :
 - Adds a 'thread holds scan lock' assertion to the LQH code
   which iterates multiple scan rows per RTB
   - Could be other useful places to add such an assertion
 - Adds a testcase running a number of MRR scans on a
   cluster with query threads

Change-Id: Ic6b725db6e70dcec7ce2b110084c2d63ad6be71b

Bug#35660890 Scans of multi ranges can lose fragment locks for second and onwards ranges

Part 3/3

Fix as proposed in bug report, remove lock release from
accScanCloseConfLab()

This avoids releasing the fragment lock early in the MRR
scan.

If this release is required for another execution sequence
then it is likely to trigger the check added in patch 1/3
if that sequence is covered by testing.

Change-Id: I6a7c7c8218e20c82254c399e02d426faf6dd9858