Bug#37004345 - MLE SP with OUT argument reports error even if error is handled inside JS
In case of MLE SP, when sql-callout query fails, the error is set on
THD. However, if the error has been handled in JS program, such
errors are ignored and are not copied to the SP caller diagnostics area.

However, before these checks are done, if there is an OUT argument in
the stored program, the previous errors on THD causes the setting of the
OUT argument in the server to fail. This causes the stored program
execution to fail which is undesirable as the previous error was
already handled in JS SP.

The fix is to clear the thd diagnostics area if the error has been
handled. This would make sure that subsequent operations like setting
the output arguments in the server doesn't fail the execution of the
stored program. The mysql_thd_attributes service has been extended
to support the operations as resetting the DA.

Change-Id: I2992a7269f48a8a6470a7bf98804a4a25b29ff13