Bug#37111452: Error selecting from a view where the view and               connection charset differs
The problem manifests with a view that is created with connection and
client character set as LATIN1, and a query on the view when connection
and client character set is UTF8. Further on, the view contains
literal values with non-ASCII characters, and the query performs
a condition pushdown into UNIONed query blocks of the view.
The problem causes zero rows to be reported when one row is expected.

The problem is related to bug#36246859, which fixed a reported error
for a similar problem. The fixed problem in this case was to properly
consider the character set of a view definition, when pushing down
conditions contained in the view to inner query blocks. But in the
test case for this bug, the view only contained ASCII characters so
the problem for bug#37111452 was overlooked.

The omission was that the condition to be pushed down was written to
a text string with wrong character set. In the test case, the string
was created with UTF8 character set, but should have been created with
LATIN1 character set. As long as literal values contain only ASCII
characters, there is no problem, but with other characters, such as
the THORN character, the resulting text will contain the UTF8
representation of the characters. When the query is finally evaluated,
comparison is against an erroneous character string and no rows match,
even if they should.

The fix to the problem is to ensure that the string is created with
the correct character set.

Change-Id: I2f4f4de791370f489368864dac61b7e1e610f69f