Bug#37235186: Unnecessary work in HashJoinIterator's constructor [4/4, error checking]
An earlier patch in this series changed the signature of
Mem_root_array::resize() so that an error code was returned if the
allocation failed.

Update the callers to check for errors and handle them.

In some cases, resize() was used only to shrink the array, so that no
allocation would ever happen. Instead of checking for allocation
errors in those cases, these callers are changed to use a more
specialized function, like chop(), pop_back() or erase(), to make it
clear that no allocation is performed.

Change-Id: Ifef747adee0ba0e3b489c95bc7f577cd8fed5c82