Bug#37248570 mysqltest_safe_process eats up empty parameter on Windows
Problem:

When you run a process in MTR test using mysqltest_safe_process
as a launcher it eats up an empty parameter you want to pass to
that process. This only affects Windows.

For example, when your code looks like that:

My::SafeProcess->run(name  => "mysqlrouter_keyring",
                       path  => $exe_mysqlrouter_keyring,
                       args  => ["account_with_auth_socket", "password", ""]);

The last parameter which is empty string, never reaches the mysqlrouter_keyring
process when it is launched.

Solution:

This patch changes the windows implementation so that the empty parameter
is quoted with ""-s and reaches the destination process.

It also fixes out-of-bound access in a check in the existing code.

Change-Id: Ic5bc4152ac59ad8992d1ce22dd3283b9584e6187