Bug#37095383  fts_delete is not crash-safe if binlog is disabled
Symptom:
When binlog is disabled, results of full-text search queries may be
incorrect after a server crash.

Cause:
In function 'fts_commit_table' a new trx object is allocated to handle
the insertion of fts_*_deleted. When binlog is disabled, an inconsistency
may occur between fulltext index and clustered index.

Fix:
Use the original trx object if it's active.

Thanks to Yin Peng at Tencent for the contribution!

Change-Id: Ib3bea350cf5f5dd3636ed105234570fe8f189fca