Bug#37243135 Ndb : Socket close error reports cause confusion
Recent refactoring of the transporter layer added the reporting
of the presence of socket shutdown errors, but not their nature.

This has caused confusion in the common case where a socket
shutdown is requested, but the socket is already closed by the
peer.

e.g.  ENOTCONN

This logging is removed.

Change-Id: Ife65079577b020879a588119c942aa756e01ad8e