Bug#37243446: Assertion: item->contextualized in Item::Item
In the patch for bug 36531144, we changed some constructors for
subquery items so that we passed an improved definition text to
their constructors. Unfortunately, the constructors with and without
POS arguments are quite different. Those with a POS argument usually
rely upon a contextualization function to add them to the THD chain
of items, as well as set the contextualized flag. Thus, some special
tests might assert because the contextualized was never set.

The fix here is to add explicit function calls to set the flag.

Change-Id: I6e836bac0beb322b1129d3edeeffc749637120e3