Bug#37263191 testDict -n TableAddAttrsUpdateMaxRecSzForLCP config change leftover
Test uses the function changeLCPInterval twice, first (as Initializer
step) to  set TimeBetweenLocalCheckpoint to 0 and second (as finalizer)
to revert it to the original value.
Due to a bug in the changeLCPInterval function, the value
restored in the finalizer is wrong, function sets a default
value instead of the original value stored in the test context
during the initializer step.

Fix: restore the correct value.

This patch also increases the TableAddAttrsUpdateMaxRecSzForLCP
max-time (in the AT test suite). Current value can be too
small in some setups.

Change-Id: I2b64678d2941b05fdf5035e3a2e3af3754e5ed14