WL#16109 Hypergraph cost model for large data (including IO cost) [2/2]
This commit adds IO cost for the basic table access paths for the
Hypergraph optimizer. This includes table scans, index scans, index
range scans and lookups. The cost of reading an n-byte object from
disk is estimated by an affine function:

C(n) = A+B*n

Note that the granularity of n is the size of storage engine blocks,
typically 16KB in case of InnoDB.

The values for constants A and B were derived from running the JET
benchmark of the previous commit in a cloud environment.

In preliminary tests using a TPC-H benchmark, the running time fell by
about 27 percent.

Change-Id: I4e5bfd77b11376a0ddde1879b720d02192143321