Bug#37228735 Timestamp flow issue - start timestamp is bigger than the last timestamp
When a Binlogging MySQL Server connected to Ndb Cluster starts it checks
for the existence of binary logs.
If binary logs already exist then it may be restarting, and so it will
write an Incident event so that any downstream replicas can detect
the potential for lost events.

The resulting Binlogging pattern is then :

  Binlog.A       Last binlog file being written previously

  Binlog.A+1     New binlog file created on server start
                 Contains Incident event

  Binlog.A+2     New binlog file created after writing
                 incident event

Binlog.A+1 normally contains a short sequence of events such as:
  1  Start
  2  Previous-GTIDs
  3  Anonymous_GTID
  4  Incident
  5  Rotate

It is expected that the event timestamp of events 2,3,4,5 should be
greater than or equal to that of event 1.

However, it was possible for events 3,4,5 to be created after events
1 and 2, but with a lower timestamp, so that the expected ordering
was not followed.

This is fixed by ensuring that a fresh timestamp is used prior to
writing an incident to the Binlog on startup.

Change-Id: Ia2de479311a92f7be774b49c906467f5c03ef926