Bug#37273525 libc++ is deprecating/removing char_traits<unsigned char> [noclose]
The LLVM libc++ library is deprecating/removing char_traits<unsigned char>
  See https://reviews.llvm.org/D157058

The deprecation comes with clang-18, removal with clang-19

This means that our code using e.g.
  std::basic_string<uchar>
  std::basic_ostringstream<unsigned char>
will no longer compile for
  clang/gcc -stdlib=libc++

We usually build with -stdlib=libstdc++
which is the standard library that comes with gcc.
The exception is MacOS, where we always use libc++

Implement our own standards-compliant my_char_traits<unsigned char>
and use that as a drop-in replacement for std::char_traits<unsigned char>

Change-Id: I3cd16503543e74dff852d8708766181807910f00