WL#16292 Add MysqldLongBlob support to NdbRecord API
In public NDBAPI a column flag NdbDictionary::RecordSpecification::MysqldLongBlob
is added to be used with method NdbDictionary::createRecord for columns
that should be represented as MySQL LONGBLOB in record, that is a 4 byte
length in little endian followed by a 8 byte native pointer to column
value. This flag can only be used for NDB columns that are Longvarbinary
or Longvarchar. Note, these MysqldLongBlob columns will not use NDB blob
handles.

Write row
=========
For writing a row with a MysqldLongBlob column an new internal function
NdbOperation::insertATTRINFOData_NdbRecord is added that takes two
source buffers that do not need to be word aligned or word sized and
copies their concatenation to ATTRINFO storage padded with zeros to
whole word if needed. This function will be used when constructing the
Longvarbinary attribute from LONGBLOB value without needing to copy
the separate length bytes and value bytes to a common temporary storage.

Read row
========
When reading a row with a MysqldLongBlob column application need to pass
some extra storage for storing the value bytes, while length and pointer
are stored in row buffer. This is done by using the new operation option
NdbOperation::OperationOptions::OO_ROW_SIDE_BUFFER and setting the
rowSideBuffer and rowSideBufferSize in the option structure. Note that
if there are several MysqldLongBlob only one OO_ROW_SIDE_BUFFER option
are used and must be big enough to be able to store all such column
values for a row.

Scan and Query rows
===================
Receive buffers and row buffers are both automatically expanded with
space for MysqldLongBuffer column data.

Change-Id: If44a23de8cef9fe4e2a7a0cb2d9e12fee250744f