Bug#37270271 testBasic -n CheckCompletedLCPStats T1 fails sporadically with timeout
CheckCompletedLCPStats test Performs an LCP and wait it to
start and finish to check if the number of records checkpointed
is as expected.

The LCP is triggered manually via dump code
DumpStateOrd::DihStartLcpImmediately, after dump the code
test creates a logevent handler and wait for events
NDB_LE_LocalCheckpointStarted and NDB_LE_LocalCheckpointCompleted
to be written to the eventlog.
Problem is that, since LCP is triggered before logevent handler
is created that could be the case that, when event handler is
started the events that test is waiting for were already written
to the event log, and test will wait 'indefinitely' for it.

Solution: Create the eventlog handler before trigger the
LCP, this way we ensure that test will see the LCP related
events.

Change-Id: Ibf0f9007c836a68d48677029ac272a9b09a80c69