Bug#36421735: mysql server 8.3.0 crashed at               Item_rollup_sum_switcher::val_decimal
Pushing a condition down which has a aggregate function
in a WHERE clause (allowed for the case in bugpage
because aggregation is on a constant), makes the
underlying derived table evaluate the aggregate function
when it should not leading to server exit.
Solution is to not pushdown conditions with aggregate
functions.

Change-Id: I128ae44dc2623c7b4e0ebb21faaf21246ea6eb80