Bug#36921175 Complex Join Predicate with Subquery Not Possible To Offload to HeatWave [ changed solution ]
In the repro case, the presence of an outer reference in a window's
PARTITION BY hinders offloading to Heatwave.

MySQL was too liberal in allowing outer references in PARTITION BY and
ORDER BY expressions.  SQL 2016, section 7.15 window clause, SR 4
says: Each <column reference> contained in the <window partition
clause> or <window order clause> of WDEF shall unambiguously reference
a column of the derived table T that is the result of TE (the <table
expression> that immediately contains the <window clause)

Solution: disallow outer references in a window's PARTITION BY and
ORDER BY expressions

Change-Id: I33d3b5bf1c2fa1ea146da854d2a434bf646def9a