Bug#37276946 mock_server's close_all_connections() does not work
The mock-server's REST API has a way to close all connections currently
opened at the mock-server.

But calling the REST API doesn't close the connection.

Background
==========

When registering the mock-server instance with the mock-server
component, the mock-server instance is added to copy of the
mock-server-instance-map which is then destroyed.

As the mock-server instance isn't properly added, lookup the instance to
close all its connections then fails as no instance is found.

Change
======

- registry the mock-server instance in the proper instance-map.

Change-Id: I7f50b7cc8ccc87ed7fef29bd65de4d82f15673bb