Bug#36991605: Assertion `false && "Inconsistent row counts for different AccessPath objects."' failed.
The hypergraph optimizer raised an assertion failure due to
inconsistent row count estimates in a query. The inconsistent
estimates were seen in a subplan that was not part of the final plan.
The subplan was invalid, so if it had been chosen, the query would
probably have returned wrong results.

The invalid plan is not considered after bug#36734772, so the
assertion failure is no longer seen in the latest version. The
underlying bug that made the optimizer accept the invalid plan as a
valid candidate, however, is still present.

For completeness, also fix the underlying bug in case it is still
possible to have queries that are affected by it.

The problem was caused by the LIMIT 1 that is added in some semijoins
that are transformed to inner joins. The LIMIT access path didn't
preserve the "parameter tables" property of its child. Since the
parameter tables are used prevent certain invalid join orders, losing
them could mean that an invalid join order is chosen. In the original
query, an index lookup for a join condition was attempted put on the
outer side of a nested loop join with the other table referenced by
the index lookup on the inner side. It has to be the other way around.

Fixed by making NewLimitOffsetAccessPath() preserve all the basic
properties of the child access path, including the parameter tables.

Change-Id: I46113e00c8dbce51cc5aec3dc2c83000ef886396