BUG#37271768 The LOCK_TYPE in PFS.metadata_locks not change after MDL upgrade or downgrade
This fix is a contribution from George Ma at Alibaba.

Thanks for the contribution.

Problem
=======

In the metadata lock code (MDL),
a metadata lock can in some cases be
upgraded or downgraded to a different LOCK_TYPE.

The performance_schema.metadata_locks table does not reflect
changes in LOCK_TYPE

Root cause
==========

The performance schema instrumentation is lacking an api
to report changed in LOCK_TYPE.

The MDL code is not instrumented for this case,
so that the performance_schema instrumentation is never updated.

Fix
===

Extend the performance_schema MDL instrumentation to add
a new API, to report metadata lock changes.

Due to ABI constraints, this new API is available in
service version 3.

Instrument the sql/mdl.cc code to call the performance_schema
instrumentation, to report changes in LOCK_TYPE,
for upgrade and downgrade of existing locks.

Change-Id: I549ef67b7d0179d54997b6bfa2f8667ecf571237