Bug#36849121: Instability in main.index_merge_innodb_hypergraph
One of the reported failures in the bug report was actually in a
different test (main.opt_hints_index_merge_hypergraph) and was seen as
this test diff:

@@ -678,8 +678,9 @@
 EXPLAIN SELECT f2 FROM t1 WHERE f2 = 2 AND f3 = 'b' AND f4 = 'h';
 EXPLAIN
 -> Intersect rows sorted by row ID  (rows=0.444)
-    -> Index range scan on t1 using idx over (f3 = 'b' AND f4 = 'h')  (rows=1)
     -> Index range scan on t1 using f2 over (f2 = 2)  (rows=128)
+    -> Index range scan on t1 using f3 over (f3 = 'b')  (rows=128)
+    -> Index range scan on t1 using f4 over (f4 = 'h')  (rows=128)

The failing EXPLAIN statement is performed just after an ALTER TABLE
ADD KEY statement, with no ANALYZE TABLE in between. The usual fix of
adding ANALYZE TABLE makes it fail less frequently, but it still fails
occasionally, even with ANALYZE TABLE. (Possibly because ALTER TABLE
ADD KEY requests a refresh of the statistics, and if the background
thread picks up the request after ANALYZE TABLE has completed, the
EXPLAIN statement may see semi-updated statistics because of
bug#21821930.)

Instead, the key is now added at CREATE TABLE time, but as an
invisible key so that it doesn't affect the part of the test that
didn't have this index before. And the ALTER TABLE ADD KEY statement
is replaced by an ALTER TABLE ALTER INDEX statement that makes the
invisible key visible. This seems to stabilize the test case.

Change-Id: Ic4570e3642d1e2015a59b350a64a4838d657ac9d