Bug#36778475 Issues with I_S.processlist due to missing synchronization
The protocol of a non-current THD was accessed fron the code path of
I_S processlist. This caused a race condition to occur which resulted
in invalid memory being accessed. The fix is to cache information being
accessed and to make sure protocol objects of non-current THDs are not
referenced from within the I_S processlist implementation.

Change-Id: I0ce332854f0bc63879da68702846c0ca52787ca3