Bug#37293786: Wrong result with QUALIFY in table-less query
Queries that didn't access any tables, did not evaluate the QUALIFY
clause, and they returned wrong results if the expression in the
QUALIFY clause evaluated to false.

Fixed by making JOIN::attach_access_paths_for_having_and_limit() also
attach an access path for the QUALIFY clause, if there is one. The
function is also renamed to make it clear that it also handles
QUALIFY.

Additionally, teach Item_singlerow_subselect::fix_fields() that it
should not flatten "SELECT (SELECT <expr1> QUALIFY <expr2>) ..." to
"SELECT <expr1> ...", as that flattening would cause wrong results if
<expr2> evaluated to false.

Additionally, teach guaranteed_one_row() that a table-less query with
a QUALIFY clause is not guaranteed to return one row, as it can return
zero rows if the row doesn't qualify. This was needed in order to get
the nullability right in scalar subqueries that had no tables and had
a QUALIFY clause.

Change-Id: I0089a4669a9e255e4216810ae2fbcd7f5c627240