WL#16478 Added handler cardinality estimation hook
This change adds a new hook to the handlerton to perform cardinality
estimation.

At JoinHypergraph construction time, this hook is used to request
cardinality estimates from the default secondary engine.

For each hypergraph Node for which the secondary engine could provide
a cardinality estimate, the value is written in `Node->cardinality`.

Change-Id: I42e742565e1b289729d434bdebd0dec4cfa3062c