Bug#37307903 Remove redundant method and lazy initialization of a member in the Tablespace class
Description:
------------
The Tablespace class has APIs is_implicit() and is_explicit() to check
if an undo tablespace is default or externally created. These methods
implementation depends on the member m_implicit. This member is lazily
initialized though in Tablespace::set_file_name().
There are two problems with using former two methods :
1. They rely on someone to remember to call  set_file_name  at all.
2. They rely on the filename string which looks rather fragile

Also, is_implicit() is not used so it could be removed.

Fix:
----
1. Removed the m_implicit  field. We now check if undo number is
   greater than flag FSP_IMPLICIT_UNDO_TABLESPACES.
2. Redefined method num() using undo::id2num()
3. Removed the is_implicit() method.
4. Redefined the is_explicit()
5. Removed  m_num  member

Change-Id: I1bba2d95cfc8c830684b610102ed7e40b4f6dd65