Bug #35981769: heap-use-after-free in Acl_cache due to a possible race condition
When ACL statements are executed the the ACL map cache is
flushed to make sure that the role flattenings are current with
the new data introduced by the ACL statement.

There was a race condition after the map is found and before it being
removed from the cache lock-free hash. If, during that time, another
thread finds the map in the lock-free hash it is going to get used
by that thread. However, the flush thread is going to delete the map
and go on since the map is not any more on the lock free hash.

Fixed by waiting for the reference count of the map to drop down to
0 before proceeding with the delete.
This is safe, since at that point the map is not already on the acl
cache hash and the reference count can only go down at this point.

Change-Id: I556973bf8a948b270ca0f6e9b25e616607ad128d