Bug#36421704: mysql server 8.3.0 crashed at               change_to_use_tmp_fields_except_sums
Optimizer tries to remove a subquery that has a
simple expression in its projection when the
subquery does not use any tables and does not
have any set operation, or an aggregate function
or a window function etc. When such a subquery
has ROLLUP, the projected expression is wrapped
in rollup group wrapper by the time the subquery
is removed. This expression is later used in the
outer query which leads to the outer query not
finding the correct item during resolving leading
to problems later when temp table fields are setup.
Solution is to not remove a subquery when it has
ROLLUP or CUBE as the result of the projected
expression could change because of their presence.

Change-Id: I8bf6e2e56a39aaa0d57c8e16acb04b7a8f2235f1