Bug#37321762:  Item_in_subselect::finalize_materialization_transform                at sql/item_subselect.cc
A subquery which is marked for materialization during resolving is not
optimized when an impossible condition is detected else where in the
query. This created problems when optimizer was trying to detect
subqueries that needed to be materialized. Since the subquery was not
optimized, trying to access the query block's join pointer resulted in
server exit.
Solution is to check if the subquery was optimized before finalizing
the materialization transform for the subquery.

Change-Id: I35077ddf9e5164233db952fb0ce2117230d9954d