Bug#37163647 SR recovery issues
Test

testSystemRestart is extended with a new test GCPSaveLagLcpSR
which exercises a multi-NG system with :
 - All nodes but one suffering GCP_SAVE lag
 - A subsequent LCP being triggered
 - A subsequent System Restart

This exposes the problem mentioned in the bug
with CopyGCIReq copying state which leads to
the SR being unrecoverable.

The symptom is that the SR does not complete.

Problem

The DIH block in the Master/President role controls
the GCP_SAVE and COPY_GCI protocols.

The GCP_SAVE protocol results in updates to each
participating node's lastCommittedGCI metadata,
and the COPY_GCI protocol propagates this information
to all nodes.

System Restart code effectively assumes that :
 - The newestRestorableGCI is the max of the stored
   per-node lastCommittedGCIs
 - The max of the lastCommittedGCIs is restorable

The recoverability-robustness of the system is improved
by moving the checks of these assumptions back from
the distributed System Restart phase to the CopyGCIREQ
propagation phase where a live President instructs each
node to write new state to disk.

Improvement 1

If there is some logic problem that could threaten
future recoverability of the cluster, it causes the
President to halt immediately.

While this can cause an immediate service outage, it
surfaces + avoids the risk of unrecoverability.

The situation where a non recoverable set of GCI info is
distributed should be rare.  However if a running cluster is
upgraded to a version containing these checks then there is a
chance that a new version President will fail as a result of
inheriting inconsistent data from the previous old version
President.

This scenario is risky as it is in precisely this situation that
it is possible that the cluster is not SR recoverable, so we do
not want to risk needing an SR.

For this reason, as part of Master GCP takeover, all nodes will
align their inherited GCI info to ensure that it does not result
in an immediate failure.

Improvement 2

The President's logic in GCP_SAVEREQ is modified to avoid
directly updating the in-memory 'SYSFILE' as part of processing
GCP_SAVECONF signals from participating nodes.

The set of nodes which sent a CONF is instead stored in
a bitmap, leaving the lastCommittedGCI values intact.

When the GCP_SAVEREQ round is complete, the bitmap is used
to update the lastCommittedGci values atomically with the
newestRestorableGCI, so that any subsequent CopyGCIREQ
invocation will propagate them together.

This avoids an intermediate CopyGCIREQ (e.g. triggered
by the start of an LCP) attempting to propagate values
which would not be recoverable.

Change-Id: Ib2d5bf9dee5ae9c05670d02488adb39678ef3ac8