Bug#37310680 Contribution: Fix buffer overflow in ndb config
Contributed by Dirkjan Bussink.

Fix buffer overflow in ndb config

In case `with_ndb_home` is set, `buf` is allocated with `PATH_MAX` and
the home is already written into the buffer.

The additional path is written using `snprintf` and it starts off at
`len`. It still can write up to `PATH_MAX` though which is wrong, since
if we already have a home written into it, we only have `PATH_MAX - len`
available in the buffer.

In practice this buffer overflow only would happen with very long paths.

Change-Id: I0571db89d92b6649f2c7d3f754efb06dfe07d3cc