BUG#37330563: Test rpl_applier_metrics_waits failing on PB2 valgrind
After the push of BUG#35206392, the test rpl_applier_metrics_waits
started to fail on valgrind.
Before BUG#35206392 there was a single wait period when a transaction
being committed by the replication applier needed to wait for the
preceding transactions' commit.
On BUG#35206392 that wait period was split in multiple small waits.
That change was not considered on the computation of the waits due
to commit order and caused tests to fail, more precisely, by
counting one transaction wait as multiple transactions wait.

In order to restore the behavior, the metric computation was moved
to embrace all multiple small waits in a single observation, so that
is means that one transaction did wait for preceding transactions'
commit.

Change-Id: Id241110c32834a9c9b9cea82d57bb98cc456eab4